+++ shopAccountTpm/models.py
@@ -51,13 +51,13 @@ class AccountExpenses(models.Model):
     TotalIncome = models.IntegerField()
     OpeningBalance = models.IntegerField()
     NetProfit= models.IntegerField()
-    GetTime=models.IntegerField()
+    GetTime=models.BigIntegerField()
 
 class LabourExpenses(models.Model):
     
     Date = models.DateField()
     Day =models.CharField(max_length=10)
-    GetTime=models.IntegerField()
+    GetTime=models.BigIntegerField()
     AssanPresent = models.IntegerField()
     AssanExpense = models.IntegerField()
     RasheedPresent = models.IntegerField()
